- now releasing menu textures when we leave the menu and enter a game level. This needs a bit more work as we seem to leak memory somewhere in the texture manager code? - now passing bytes per pixel to texture loading functions - added a pink texture we can set when a proper texture was unable to be loaded correctly. Like how Quake does it. - fixed up a lot of xbox rendering specific code. The game will load now but things aren't rendering correctly yet.
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@388

+++ 3dc/win95/VideoModes.cpp
@@ -26,7 +26,7 @@ char *GetVideoModeDescription3()
 
 	static char buf[bufferSize + 1]; // always have room for a null character
 
-	uint32_t sizeofText = R_GetVideoModeDescription().size();
+	size_t sizeofText = R_GetVideoModeDescription().size();
 
 	// ensure we don't overflow our char buffer
 	if (sizeofText > bufferSize)
